requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2023.01.10=haa95532_0
certifi=2023.5.7=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py39h21b164f_0
cuda=11.6.1=0
cuda-cccl=11.6.55=0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=0
cuda-cudart-dev=11.6.55=0
cuda-cuobjdump=11.6.124=0
cuda-cupti=11.6.124=0
cuda-cuxxfilt=11.6.124=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight-compute=12.1.1=0
cuda-nvcc=11.6.124=0
cuda-nvdisasm=12.1.105=0
cuda-nvml-dev=11.6.55=0
cuda-nvprof=12.1.105=0
cuda-nvprune=11.6.124=0
cuda-nvrtc=11.6.124=0
cuda-nvrtc-dev=11.6.124=0
cuda-nvtx=11.6.124=0
cuda-nvvp=12.1.105=0
cuda-runtime=11.6.1=0
cuda-sanitizer-api=12.1.105=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-visual-tools=11.6.1=0
cycler=0.11.0=pypi_0
fonttools=4.39.4=pypi_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
idna=3.4=py39haa95532_0
imageio=2.28.1=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2023.1.0=h59b6b97_46319
joblib=1.2.0=pypi_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=pypi_0
lazy-loader=0.2=pypi_0
lerc=3.0=hd77b12b_0
libcublas=11.9.2.110=0
libcublas-dev=11.9.2.110=0
libcufft=10.7.1.112=0
libcufft-dev=10.7.1.112=0
libcurand=10.3.2.106=0
libcurand-dev=10.3.2.106=0
libcusolver=11.3.4.124=0
libcusolver-dev=11.3.4.124=0
libcusparse=11.7.2.124=0
libcusparse-dev=11.7.2.124=0
libdeflate=1.17=h2bbff1b_0
libnpp=11.6.3.124=0
libnpp-dev=11.6.3.124=0
libnvjpeg=11.6.2.124=0
libnvjpeg-dev=11.6.2.124=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
matplotlib=3.7.1=pypi_0
mkl=2023.1.0=h8bd8f75_46356
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.6=py39hf11a4ad_1
mkl_random=1.2.2=py39hf11a4ad_1
networkx=3.1=pypi_0
nsight-compute=2023.1.1.4=0
numpy=1.24.3=py39h79a8e48_1
numpy-base=1.24.3=py39h8a87ada_1
openssl=1.1.1t=h2bbff1b_0
packaging=23.1=pypi_0
pillow=9.4.0=py39hd77b12b_0
pip=23.0.1=py39haa95532_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py39haa95532_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.16=h6244533_2
python-dateutil=2.8.2=pypi_0
pytorch=1.13.1=py3.9_cuda11.6_cudnn8_0
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=pypi_0
requests=2.29.0=py39haa95532_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.9.1=pypi_0
setuptools=66.0.0=py39haa95532_0
six=1.16.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.4.12=pypi_0
tk=8.6.12=h2bbff1b_0
torchaudio=0.13.1=pypi_0
torchvision=0.14.1=pypi_0
typing_extensions=4.5.0=py39haa95532_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
xz=5.4.2=h8cc25b3_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
